backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-genai>=0.8.0
pydantic==2.8.2
python-dotenv==1.0.0
requests==2.31.0
lunardate==0.2.2